improve docstring for show_rectangles
I think that "the x postion" of a rectangle is ambiguous. In this case it refers to the center of the retangle, but in many other cases in matplotlib it would be the lower left corner.
So I added the words "the x position of the center of the rectangle".
For symmetry, I introduced the same change for circles and ellipses, although it's unambiguous in that case anyway.